+++ sdk/ai/azure-ai-projects/tests/test_datasets.py
@@ -89,9 +89,9 @@ class TestDatasets(TestBase):
             )
 
             print(f"[test_datasets_upload_file] Get credentials of an existing Dataset version `{dataset_version}`:")
-            asset_credential = project_client.datasets.get_credentials(name=dataset_name, version=str(dataset_version))
-            print(asset_credential)
-            TestBase.validate_asset_credential(asset_credential)
+            dataset_credential = project_client.datasets.get_credentials(name=dataset_name, version=str(dataset_version))
+            print(dataset_credential)
+            TestBase.validate_dataset_credential(dataset_credential)
 
             """
             print("[test_datasets_upload_file] List latest versions of all Datasets:")
@@ -185,9 +185,9 @@ class TestDatasets(TestBase):
             )
 
             print(f"[test_datasets_upload_file] Get credentials of an existing Dataset version `{dataset_version}`:")
-            asset_credential = project_client.datasets.get_credentials(name=dataset_name, version=str(dataset_version))
-            print(asset_credential)
-            TestBase.validate_asset_credential(asset_credential)
+            dataset_credential = project_client.datasets.get_credentials(name=dataset_name, version=str(dataset_version))
+            print(dataset_credential)
+            TestBase.validate_dataset_credential(dataset_credential)
 
             print(
                 f"[test_datasets_upload_file] Delete Dataset `{dataset_name}`, version `{dataset_version}` that was created above."